combine tests that share to reader and writer
