+++ e2e/app/page/workspace-about-page.ts
@@ -72,10 +72,11 @@ export default class WorkspaceAboutPage extends WorkspaceBase {
   }
 
   // if the collaborator is already on this workspace, just remove them before continuing.
-  async removeCollaborator(name: string): Promise<void> {
+  async removeCollaborator(email: string): Promise<void> {
     const modal = await this.openShareModal();
-    await modal.removeUser(name);
+    await modal.removeUser(email);
     await waitWhileLoading(this.page);
+    await this.waitForLoad();
   }
 
   async getAboutLockedWorkspaceIcon(): Promise<void> {
